+++ src/main/java/gift/auth/Interceptor.java
@@ -1,4 +1,4 @@
-package gift.authService;
+package gift.auth;
 
 import gift.exception.AuthException;
 import io.jsonwebtoken.Claims;
